App: Adjust code/tests to app/ contained codebase
