+++ app/tests/run_tests.py
@@ -2,18 +2,17 @@
 # -*- coding: utf-8 -*-
 
 import os
-import sys
-
 import unittest
 import coverage
 
-sys.path.append(os.path.abspath(os.path.join(
-    os.path.dirname(__file__), 'app'
-    )))
+appDir = os.path.abspath(os.path.join(
+    os.path.dirname(__file__),
+    '..'
+    ))
 
 loader = unittest.TestLoader()
-suite = loader.discover('tests')
-cov = coverage.coverage(branch=True, source=['app'])
+suite = loader.discover(os.path.join(appDir, 'tests'))
+cov = coverage.coverage(branch=True, source=[appDir])
 
 runner = unittest.TextTestRunner(verbosity=2)
 cov.start()
@@ -22,5 +21,5 @@ if result.wasSuccessful():
     cov.stop()
     print('Coverage Summary:')
     cov.report(show_missing=True)
-    cov.html_report(directory='coverage/app')
+    cov.html_report(directory='coverage')
     cov.erase()
