+++ app/tests/test_utils.py
@@ -6,7 +6,7 @@ sys.path.append(os.path.abspath(os.path.join(
     os.path.dirname(__file__), '..'
     )))
 
-from app.utils import markdownToToc
+from utils import markdownToToc
 
 class UtilsTestCase(unittest.TestCase):
